change url

+++ user_career_more/lib/service/models/service_model.dart
@@ -157,7 +157,7 @@ class ServiceModel implements Decodable{
     expertId = json['expert_id'];
     scheduleTime = DateTime.parse(json['schedule_time']);
     duration = json['duration'];
-    totalPrice = json['total_price'];
+    totalPrice = (json['total_price']).toInt();
     noteMessage = json['note_message'];
     contactMethod = json['contact_method'];
     locationName = json['location_name'];
